### 2025-06-06 #8362 @editorialbot
[PRE REVIEW]: PySGN: A Python package for constructing synthetic geospatial networks
**Submitting author:** @wang-boyu (<a href="http://orcid.org/0000-0001-9879-2138">Boyu Wang</a>)
**Repository:** https://github.com/wang-boyu/pysgn
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.1
