### 2025-08-01 22c4a66
fix joss paper bibtex entries
