+++ paper/paper.bib
@@ -46,20 +46,19 @@ year={1960}
   pages={362--390},
   year={2017},
   publisher={Springer},
-  doi={10.1007/s10588-016-9232-2},
+  doi={10.1007/s10588-016-9232-2}
 }
 
-@article{osti_960616,
-title = {Exploring network structure, dynamics, and function using NetworkX},
-author = {Hagberg, Aric and Swart, Pieter J. and Schult, Daniel A.},
-abstractNote = {NetworkX is a Python language package for exploration and analysis of networks and network algorithms. The core package provides data structures for representing many types of networks, or graphs, including simple graphs, directed graphs, and graphs with parallel edges and self loops. The nodes in NetworkX graphs can be any (hashable) Python object and edges can contain arbitrary data; this flexibility mades NetworkX ideal for representing networks found in many different scientific fields. In addition to the basic data structures many graph algorithms are implemented for calculating network properties and structure measures: shortest paths, betweenness centrality, clustering, and degree distribution and many more. NetworkX can read and write various graph formats for eash exchange with existing data, and provides generators for many classic graphs and popular graph models, such as the Erdoes-Renyi, Small World, and Barabasi-Albert models, are included. The ease-of-use and flexibility of the Python programming language together with connection to the SciPy tools make NetworkX a powerful tool for scientific computations. We discuss some of our recent work studying synchronization of coupled oscillators to demonstrate how NetworkX enables research in the field of computational networks.},
-doi = {},
-url = {https://www.osti.gov/biblio/960616}, journal = {},
-number = ,
-volume = ,
-place = {United States},
-year = {2008},
-month = {1}
+@conference{osti_960616,
+  author       = {Hagberg, Aric and Swart, Pieter J. and Schult, Daniel A.},
+  title        = {Exploring network structure, dynamics, and function using NetworkX},
+  annote       = {NetworkX is a Python language package for exploration and analysis of networks and network algorithms. The core package provides data structures for representing many types of networks, or graphs, including simple graphs, directed graphs, and graphs with parallel edges and self loops. The nodes in NetworkX graphs can be any (hashable) Python object and edges can contain arbitrary data; this flexibility mades NetworkX ideal for representing networks found in many different scientific fields. In addition to the basic data structures many graph algorithms are implemented for calculating network properties and structure measures: shortest paths, betweenness centrality, clustering, and degree distribution and many more. NetworkX can read and write various graph formats for eash exchange with existing data, and provides generators for many classic graphs and popular graph models, such as the Erdoes-Renyi, Small World, and Barabasi-Albert models, are included. The ease-of-use and flexibility of the Python programming language together with connection to the SciPy tools make NetworkX a powerful tool for scientific computations. We discuss some of our recent work studying synchronization of coupled oscillators to demonstrate how NetworkX enables research in the field of computational networks.},
+  url          = {https://www.osti.gov/biblio/960616},
+  place        = {United States},
+  organization = {Los Alamos National Laboratory (LANL)},
+  year         = {2007},
+  month        = {12},
+  doi          = {10.25080/tcwv9851}
 }
 
 @software{joris_van_den_bossche_2024_12625316,
