### 2025-11-02 #8362 @wang-boyu
Hi @jGaboardi thank you for taking our paper!

> can you comment on the role of your 3 co-authors for this submission

Their contributions are mainly on the algorithms or methods themselves, whereas I mostly worked on its Python implementation and optimization. So their contributions are not directly visible through code.

Not sure whether this is how it's supposed to work, but here are the five reviewers I found in the reviewer database who may be faimilar with this paper's topic:

- cbueth
- abhishektiwari
- benyamindsmith
- mszell
- simonprovost
