### 2025-11-02 04b717d
fix ruff errors
