### 2025-11-05 #9346 @tomalrussell
- Queried what contributions other authors made - noted they contributed ideas behind the algorithms implemented in PySGN in https://github.com/wang-boyu/pysgn/issues/13
    - This submission may be slightly on the light side, given the lines of code in the main package (688 according to cloc, JOSS would flag under 1000) and a rough estimate from commit history suggests the 39 commits (excluding dependabot) fall in 2-3months effort over the period since Oct 2024.
    - It does make random spatial graph generation considerably simpler, with three core methods for generating edges given a pre-specified set of nodes
    - https://github.com/wang-boyu/pysgn/issues/17 - noted as nice-to-have but not essential, agreed
    - https://github.com/wang-boyu/pysgn/issues/18 - implemented in https://github.com/wang-boyu/pysgn/pull/33 pending merge
    - https://github.com/wang-boyu/pysgn/issues/21 - argued that node synthesis is out of scope

  - README examples raise errors, pending https://github.com/wang-boyu/pysgn/pull/24 and release

   - dependencies listed in `pyproject.toml`, will be improved by https://github.com/wang-boyu/pysgn/pull/28 to help conda-environment users and when https://github.com/wang-boyu/pysgn/pull/24 is merged and a new version of pysgn is released
  - examples in README will be improved - see https://github.com/wang-boyu/pysgn/issues/12 and https://github.com/wang-boyu/pysgn/issues/16 - pending merge and release as above
  - improved through https://github.com/wang-boyu/pysgn/issues/20 and https://github.com/wang-boyu/pysgn/pull/23/

  - improved through https://github.com/wang-boyu/pysgn/issues/14 and https://github.com/wang-boyu/pysgn/pull/31
  - Extended through https://github.com/wang-boyu/pysgn/issues/22 and https://github.com/wang-boyu/pysgn/pull/31
