### 2025-11-08 wang-boyu/pysgn#11 @gegen07
Example usage for pysgn
Following openjournals/joss-reviews#9346 requirements, I would like to see `pysgn` examples notebooks to review your work with regard to performance and usability properly. It would also be nice for users to test your functionalities easily.
