### 2025-11-21 wang-boyu/pysgn#12 @tomalrussell
README examples raise errors as written
Hi @wang-boyu - thanks for submitting this package to JOSS, it looks good. I'm going to raise a few issues but all should be fixable.

The code snippets in the README (and docs front page) refer to `'path/to/your/geospatial_data.shp'` without specifying what that should contain.

1. I'd suggest using a `GeoDataFrame` literal here, or a small example dataset e.g. from [geodatasets](https://pypi.org/project/geodatasets/) as geopandas uses for some tutorials.
1. Note any restrictions on geometry types - as I understand it, the methods accept either all `Point` or all `Polygon`

https://github.com/wang-boyu/pysgn/blob/6945f88499386d3a252c20e92e56da621dd585cb/README.md?plain=1#L29

```python
import geodatasets
# ...
gdf = gpd.read_file(geodatasets.get_path("geoda groceries")).explode()  # explode to go from Multi- to Points
```

Part of https://github.com/openjournals/joss-reviews/issues/9346
