### 2025-11-21 wang-boyu/pysgn#13 @tomalrussell
Clarify author contributions
Could you clarify the author contributions?

As far as I can see, @wang-boyu has authored all of the code. What were Crooks, Anderson, and Züfle's contributions?

Part of https://github.com/openjournals/joss-reviews/issues/9346
