### 2025-11-21 wang-boyu/pysgn#14 @tomalrussell
Illustrative example in paper
The summary section: paragraph 1 introduces the area clearly, but doesn't motivate or explain the Erdos-Renyi/Watts-Strogatz/Barabasi-Albert models. Paragraph 2 dives into technical options for use of the package, might be better later in the paper.

An illustrative example (with figure) in the paper would help explain purpose of the software for a diverse, non-specialist audience.

[Züfle (2024)](https://doi.org/10.1145/3672557) looks like a rich source of examples - or consider drawing on something like [Gastner and Newman (2006)](https://doi.org/10.1140/epjb/e2006-00046-8) to e.g. illustrate the qualities of some empirical network and create a related synthetic network.

Part of https://github.com/openjournals/joss-reviews/issues/9346
