### 2025-11-21 wang-boyu/pysgn#16 @tomalrussell
Projected CRS usage in examples
Suggest using a more concrete example for the getting started guide. Web Mercator is not a particularly accurate CRS for scale and location (see [remarks](https://epsg.io/3857) on EPSG), and points centred around (0, 0) are in the ocean around the Gulf of Guinea.

Consider locating the example somewhere specific, with appropriate projected CRS, e.g. perhaps:
- EPSG:3629 for New York West - is this good locally for you at Buffalo?
- EPSG:5072 for ConUS (Albers equal area, metric)

To support this, if desirable - could also use contextily or cartopy to add a background map for context in example plots.

Part of https://github.com/openjournals/joss-reviews/issues/9346
