### 2025-11-21 wang-boyu/pysgn#17 @tomalrussell
Allow geographic distance calculations
**What's the problem this feature will solve?**
Distance calculations currently assume a projected coordinate system. This requires users to choose a CRS, if their data source is not projected, and limits the scale of applicability.

For example, a synthetic network of global flight connections might naturally use a great-circle or geodetic distance to generate links between airport point locations given in lat/lon, including edges crossing the antimeridian.

**Describe the solution you'd like**
Option to specify distance measure using geodesic calculation or great-circle approximation for geographic coordinates.

**Additional context**
In general, distance calculations are done directly with pairs of coordinates, so could switch out implementation fairly straightforwardly:

https://github.com/wang-boyu/pysgn/blob/6945f88499386d3a252c20e92e56da621dd585cb/pysgn/geo_barabasi_albert_network.py#L175

https://github.com/wang-boyu/pysgn/blob/6945f88499386d3a252c20e92e56da621dd585cb/pysgn/geo_erdos_renyi_network.py#L140-L142

See for example
- [pyproj.Geod.line_length](https://pyproj4.github.io/pyproj/stable/api/geod.html#pyproj.Geod.line_length)
- [haversine](https://github.com/mapado/haversine)

Nearest-neighbour calculations rely on a KDTree - depending on performance requirements, might need to look into something like [`S2PointIndex`](http://s2geometry.io/devguide/cpp/quickstart.html) for similar indexed lookups on geographic coordinates:

https://github.com/wang-boyu/pysgn/blob/6945f88499386d3a252c20e92e56da621dd585cb/pysgn/geo_watts_strogatz_network.py#L63
