### 2025-11-21 wang-boyu/pysgn#17 @tomalrussell
(Just to clarify, I think this would be nice to have, but would not be essential for JOSS review)
