### 2025-11-21 wang-boyu/pysgn#18 @tomalrussell
Export graphs to GeoDataFrames, save to file
**What's the problem this feature will solve?**
Having created synthetic spatial networks, it can be useful to save or export these in standard geospatial formats (on disk) or as `GeoDataFrame`s for further spatial analysis.

**Describe the solution you'd like**
Methods called something like `graph_to_gdfs`, and perhaps also `graph_to_file`
- nodes as a `GeoDataFrame` with Points and `node_id` column
- edges as a `GeoDataFrame` with LineStrings and `u_id`, `v_id` columns (or other naming convention for the two node ids connected by an edge)
- include any attributes as additional columns
- `graph_to_file` could save to geopackage with `nodes` and `edges` layers by default

**Additional context**
This would support general compatibility with the spatial analysis ecosystem in Python and other languages.
