### 2025-11-21 wang-boyu/pysgn#19 @tomalrussell
Optional dependencies for getting started
Additional dependencies, as well as `pysgn` and its non-optional dependences, are required to run the getting started notebook:

```
pip install pysgn matplotlib notebook
jupyter notebook docs/getting_started.ipynb
```

Part of https://github.com/openjournals/joss-reviews/issues/9346
