### 2025-11-21 wang-boyu/pysgn#20 @tomalrussell
Improve test coverage of edge/error cases
Test coverage is high overall, and I wouldn't insist on 100% coverage as a reasonable goal in and of itself, but it would be worth throwing some more cases at the library.

In particular, triggering edge cases, errors and warnings - coverage indicates "raise ValueError" and "warnings.warn" cases are not tested - e.g. in https://app.codecov.io/gh/wang-boyu/pysgn/blob/main/pysgn%2Fgeo_barabasi_albert_network.py

For example:
- [x] test `node_order` when specified by a column name string (and expect error if string is not in gdf.columns)
- [x] test error is raised by specifying m > number of nodes
- [x] test corner case that triggers "no edges created" warning

Part of https://github.com/openjournals/joss-reviews/issues/9346
