### 2025-11-21 wang-boyu/pysgn#21 @tomalrussell
Node synthesis
**What's the problem this feature will solve?**
The core functions of the package create edges between a provided set of nodes. Examples in the documentation demonstrate uniform random sampling within a 2D bounding box to create a set of nodes. In general, to create complete synthetic spatial networks, there is a need to synthesise nodes with locations.

**Describe the solution you'd like**
One or more methods to generate sets of nodes to seed complete network creation:
- as in the docs, uniform random sampling within a bounding box

Perhaps also:
- sampling masked by a polygon (e.g. to sample points over land)
- sampling weighted by a distribution (e.g. according to population density)
- sampling uniformly on the sphere
