### 2025-11-21 wang-boyu/pysgn#22 @tomalrussell
Paper: relation to other work
The paper should reference other commonly-used packages and make clear relation to other work.

In Python spatial networks packages, there are several in the space of handling and processing spatial network data, including:
- pysal/spaghetti
- uscuni/neatnet
- gboeing/osmnx

In network synthesis, both [networkx](https://networkx.org/documentation/stable/tutorial.html#using-a-stochastic-graph-generator-e-g) and [igraph](https://python.igraph.org/en/stable/generation.html#random-graphs) provide (non-spatial) random graph generators.

Part of https://github.com/openjournals/joss-reviews/issues/9346
