### 2025-11-21 wang-boyu/pysgn#13 @wang-boyu
Other authors mainly contributed to the original ideas of the algorithms, generated before the current repository. This `pysgn` package is a Python implementation of the algorithms. Hope this clarifies!
