### 2025-11-21 c8aaeb0
update links and installation instructions in README.md
