### 2025-11-22 5180d4c
add tests for warnings and value errors
