+++ .github/workflows/build.yml
@@ -47,7 +47,7 @@ jobs:
       uses: actions/cache@v4
       with:
         path: ~/.cache/pip
-        key: ${{ runner.os }}-pip-${{ hashFiles('**/pyproject.toml') }}
+        key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml', '*/pyproject.toml') }}
         restore-keys: |
           ${{ runner.os }}-pip-
 
