### 2025-11-22 de0efba
use data from geodatasets in examples
