### 2025-11-23 wang-boyu/pysgn#11 @gegen07
Re-opening to raise another question.

(1) I think it will be clearer to move these example notebooks to a folder named `examples/` since the `docs/` folder has no related code with examples.
(2) Besides the excellent getting-started notebook, it would be good to have a real-world example, as mentioned in paper "can be used for various simulation and analysis tasks". Could you show an example of how pysgn improves these analyses?
