### 2025-11-23 wang-boyu/pysgn#25 @gegen07
Conda Support
Although PySGN is already packaged on PyPI, some researchers and developers manage their environment with Anaconda-like package manager. Is there any reason not to create an `environment.yml` and upload it to Conda-forge?

Part of https://github.com/openjournals/joss-reviews/issues/9346.
