### 2025-11-23 wang-boyu/pysgn#26 @gegen07
Performance Report
Happy to see a report on the `getting-started` notebook, but I have two questions related to it.

(1) Are there any improvements in the `networkx` random graph generation method regarding memory and CPU usage?

Note: I think it would be interesting to see comparisons between the methods in the paper, if there are any differences.

(2) Although I know that evaluating Big-O notation isn't considered important in contexts other than Computer Science, it would be valuable for the package to compare the asymptotic complexity of these three methods and discuss in detail where, how, and when each method excels.

Note: In `getting-started` notebook, the Erdös-Rényi graph seems to be more computationally intensive than other methods. Try to be clearer about which tasks it excels the other random-graph generation methods (Barabási-Albert and Watts-Strogatz).

Part of https://github.com/openjournals/joss-reviews/issues/9346.
