### 2025-11-23 wang-boyu/pysgn#12 @wang-boyu
Updated readme file to use geodatasets in the code examples now, as part of https://github.com/wang-boyu/pysgn/pull/24. The updated readme now looks like this: https://github.com/wang-boyu/pysgn/blob/docs/use-geodatasets/README.md
