### 2025-12-23 fcc6030
add related work and example sections
