+++ paper/paper.md
@@ -37,15 +37,124 @@ doi: 10.3847/xxxxx
 
 # Summary
 
-PySGN (**Py**thon for **S**ynthetic **G**eospatial **N**etworks) is an open-source Python package designed to construct synthetic undirected geospatial networks that consider the coordinates of nodes. By incorporating the spatial information, the package extends classic network generation models, such as Erdős-Rényi [@erdos1960evolution], Watts-Strogatz [@watts1998collective] and Barabási–Albert [@barabasi1999emergence]. It creates synthetic geospatial networks that mimic the spatial relationships observed in real-world networks (e.g., decaying connectivity with respect to node distance) and can be used for various simulation and analysis tasks. PySGN integrates with the PyData ecosystem, utilizing libraries, such as GeoPandas [@joris_van_den_bossche_2024_12625316] and NetworkX [@osti_960616], to facilitate the creation, utilization, and analysis of synthetic geospatial networks.
+Synthetic networks are commonly used to study the structure and dynamics of social systems, transportation infrastructure and other complex phenomena. Classical random graph models, such as the Erdős-Rényi, Watts-Strogatz and Barabási-Albert models, generate abstract networks with different structural characteristics: the Erdős-Rényi model connects nodes at random with equal probability, the Watts-Strogatz model rewires a ring lattice to produce small‑world networks, and the Barabási-Albert model yields scale‑free networks through preferential attachment [@erdos1960evolution; @watts1998collective; @barabasi1999emergence]. In their standard form these models ignore the spatial positions of nodes; yet in many empirical settings (e.g., human social networks, commuting patterns or infrastructure networks) proximity strongly influences who connects to whom.
 
-PySGN offers flexibility in parametrization, allowing users to define networks by an average node degree or expected degree for each node. Users can also define custom constraints and control the rewiring probability, distance decay, and other key parameters, making it adaptable to various research contexts. PySGN is intended for researchers and practitioners in fields, such as urban planning, epidemiology, and social science, who require robust tools for simulating and analyzing complex geospatial networks.
+PySGN (**Py**thon for **S**ynthetic **G**eospatial **N**etworks) is an open‑source Python package that extends the classical models to geospatial contexts. It embeds nodes in geographic coordinate space, modifies connection rules to decay with distance, and allows users to incorporate clustering and preferential attachment while respecting spatial constraints. By combining these ingredients, PySGN generates synthetic geosocial networks that mimic the spatial relationships observed in real‑world networks. The package integrates with the PyData ecosystem through libraries like GeoPandas [@joris_van_den_bossche_2024_12625316] and NetworkX [@osti_960616] and provides a flexible interface for research and simulation.
+
+With PySGN, users can specify parameters such as average degree and decay function among other options, and construct a geospatially embedded network as NetworkX graph objects. They can also fine‑tune the generation process by defining custom distance functions or constraints. The resulting networks can be further analyzed and visualized thereafter. The package is intended for researchers and practitioners in fields such as urban planning, epidemiology, infrastructure resilience and social science who require robust tools for simulating and analyzing complex geospatial networks.
 
 # Statement of Need
 
[…]
