+++ paper/paper.bib
@@ -1,33 +1,33 @@
 @inproceedings{10.1145/3615896.3628345,
-author = {Gallagher, Ketevan and Anderson, Taylor and Crooks, Andrew and Z\"{u}fle, Andreas},
-title = {Synthetic Geosocial Network Generation},
-year = {2023},
-isbn = {9798400703584},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/3615896.3628345},
-doi = {10.1145/3615896.3628345},
-abstract = {Generating synthetic social networks is an important task for many problems that study humans, their behavior, and their interactions. Geosocial networks enrich social networks with location information. Commonly used models to generate synthetic social networks include the classical Erd\H{o}s-R\'{e}nyi, Barab\'{a}si-Albert, and Watts-Strogatz models. However, these classic social network models do not consider the location of individuals. Real-world geosocial networks do exhibit a strong spatial autocorrelation, thus having a higher likelihood of a social connection between agents that are spatially close. As such, recent variants of the three classical models have been proposed to consider location information. Yet, these existing solutions assume that individuals are located on a uniform lattice and exhibit certain limitations when applied to real-world data that exhibits clusters. In this work, we discuss these limitations and propose new approaches to extend the three classic social network generation models to geosocial networks. Our experiments show that our generated synthetic geosocial networks address the shortcomings of the state-of-the-art models and generate realistic geosocial networks that exhibit high similarity to real-world geosocial networks.},
-booktitle = {Proceedings of the 7th ACM SIGSPATIAL Workshop on Location-Based Recommendations, Geosocial Networks and Geoadvertising},
-pages = {15–24},
-numpages = {10},
-keywords = {geosocial networks, network generation, synthetic social networks, erdos-renyi, watts-strogatz, barabasi-albert},
-location = {Hamburg, Germany},
-series = {LocalRec '23}
-}
+  author = {Gallagher, Ketevan and Anderson, Taylor and Crooks, Andrew and Z\"{u}fle, Andreas},
+  title = {Synthetic Geosocial Network Generation},
+  year = {2023},
+  isbn = {9798400703584},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/3615896.3628345},
+  doi = {10.1145/3615896.3628345},
+  booktitle = {Proceedings of the 7th ACM SIGSPATIAL Workshop on Location-Based Recommendations, Geosocial Networks and Geoadvertising},
+  pages = {15–24},
+  numpages = {10},
+  keywords = {geosocial networks, network generation, synthetic social 
+networks, erdos-renyi, watts-strogatz, barabasi-albert},
+  location = {Hamburg, Germany},
+  series = {LocalRec '23}
+  }
 
-@inbook{erdos1960evolution,
-title = {On the evolution of random graphs},
-author = {P. Erdös and A. Rényi},
-journal={Publ. math. inst. hung. acad. sci},
-volume={5},
-number={1},
-pages={17--60},
-year={1960}
-}
+  @inbook{erdos1960evolution,
+  title = {On the Evolution of Random Graphs},
+  author = {P. Erd\"os and A. R\'enyi},
+  journal={Publ. math. inst. hung. acad. sci},
+  volume={5},
+  number={1},
+  pages={17--60},
+  year={1960}
+  }
 
 @article{watts1998collective,
[…]
