+++ .github/workflows/build.yml
@@ -44,7 +44,7 @@ jobs:
         uv pip install --system .[dev]
 
     - name: Cache dependencies
-      uses: actions/cache@v4
+      uses: actions/cache@v5
       with:
         path: ~/.cache/pip
         key: ${{ runner.os }}-pip-${{ hashFiles('pyproject.toml', '*/pyproject.toml') }}
