+++ .github/workflows/release.yml
@@ -37,7 +37,7 @@ jobs:
         run: python -m build
       - name: Upload package as artifact to GitHub
         if: github.repository == 'wang-boyu/pysgn' && startsWith(github.ref, 'refs/tags')
-        uses: actions/upload-artifact@v5
+        uses: actions/upload-artifact@v6
         with:
           name: package
           path: dist/
