### 2026-01-08 2b09732
avoid redundant id col as node attribute when used as key
