### 2026-01-08 d1cc7b7
avoid redundant id col as node attribute when used as key
