### 2026-01-08 25d4775
store crs and id_col in graph
