### 2026-01-08 3028fa9
store crs and id_col in graph
