+++ tests/test_geo_barabasi_albert_network.py
@@ -134,6 +134,31 @@ def test_geographic_warning(point_gdf: gpd.GeoDataFrame) -> None:
         geo_barabasi_albert_network(geographic_gdf, m=2)
 
 
+def test_stores_crs_and_default_id(point_gdf: gpd.GeoDataFrame) -> None:
+    """Graph should persist CRS and default id metadata."""
+    g = geo_barabasi_albert_network(point_gdf, m=1, scaling_factor=1e-3)
+
+    assert g.graph["crs"] == point_gdf.crs
+    assert g.graph["id_col"] == "index"
+
+
+def test_stores_custom_id(point_gdf: gpd.GeoDataFrame) -> None:
+    """Graph should store a custom identifier column name."""
+    g = geo_barabasi_albert_network(point_gdf, m=1, id_col="id", scaling_factor=1e-3)
+
+    assert g.graph["id_col"] == "id"
+
+
+def test_stores_none_crs(point_gdf: gpd.GeoDataFrame) -> None:
+    """Graph should explicitly record a missing CRS and warn."""
+    gdf_no_crs = point_gdf.set_crs(None, allow_override=True)
+    with pytest.warns(UserWarning):
+        g = geo_barabasi_albert_network(gdf_no_crs, m=1, scaling_factor=1e-3)
+
+    assert "crs" in g.graph
+    assert g.graph["crs"] is None
+
+
 def test_index_as_id(point_gdf: gpd.GeoDataFrame) -> None:
     """Test network creation using DataFrame index as node IDs.
 
