### 2026-01-08 ba0ac1d
store crs and id_col in graph
