### 2026-01-08 92ada1a
implement graph_to_gdf to export graph to nodes and edges
