+++ pysgn/geo_watts_strogatz_network.py
@@ -6,12 +6,16 @@ from collections.abc import Callable
 import geopandas as gpd
 import networkx as nx
 import numpy as np
-import pandas as pd
 from loguru import logger
 from sklearn.neighbors import KDTree
 from tqdm.auto import tqdm
 
-from .utils import _create_k_col, _find_scaling_factor, _set_node_attributes
+from .utils import (
+    _create_k_col,
+    _find_scaling_factor,
+    _get_id_col_array,
+    _set_node_attributes,
+)
 
 
 def _get_nearest_nodes(
@@ -244,12 +248,7 @@ def geo_watts_strogatz_network(
         raise ValueError("k must be greater than 0")
     if not 0 <= p <= 1:
         raise ValueError("p must be between 0 and 1")
-    if id_col is not None:
-        if id_col == "index" and isinstance(gdf.index, pd.MultiIndex):
-            raise ValueError("Multi-index is not supported")
-        id_col_array = gdf.index.values if id_col == "index" else gdf[id_col].values
-        if len(np.unique(id_col_array)) != len(id_col_array):
-            raise ValueError("ID column must contain unique values")
+    id_col_array = _get_id_col_array(gdf, id_col)
     if isinstance(k, int | float | str):
         nearest_neighbors, degree_centrality_array = _get_nearest_nodes(
             gdf,
@@ -265,7 +264,11 @@ def geo_watts_strogatz_network(
     rewire_count = 0
     graph = nx.Graph()
     graph.graph["crs"] = gdf.crs
-    graph.graph["id_col"] = id_col if id_col is not None else "index"
+    if id_col == "index":
+        graph.graph["id_col"] = "index"
+        graph.graph["index_name"] = gdf.index.name
+    else:
+        graph.graph["id_col"] = id_col
     # use centroid if geometry is a polygon
     if gdf.geometry.geom_type.iloc[0] == "Polygon":
         pos_x_array = gdf.geometry.centroid.x.values
