### 2026-01-08 1cdea6a
implement graph_to_gdf to export graph to nodes and edges
