### 2026-01-08 b245d2c
implement graph_to_gdf to export graph to nodes and edges
