+++ tests/test_utils.py
@@ -1,10 +1,18 @@
+import re
+
 import geopandas as gpd
 import networkx as nx
 import numpy as np
+import pandas as pd
 import pytest
-from shapely.geometry import Point
+from shapely.geometry import LineString, Point, Polygon
 
-from pysgn.utils import _create_k_col, _set_node_attributes
+from pysgn.utils import (
+    _create_k_col,
+    _get_id_col_array,
+    _set_node_attributes,
+    graph_to_gdf,
+)
 
 
 def test_create_k_col_even_integer():
@@ -60,6 +68,51 @@ def test_create_k_col_reproducibility():
     )
 
 
+def test_get_id_col_array_defaults_to_positional_index():
+    """_get_id_col_array should return positional ids when id_col=None."""
+    gdf = gpd.GeoDataFrame({"id": [10, 11], "geometry": [Point(0, 0), Point(1, 1)]})
+
+    result = _get_id_col_array(gdf, None)
+
+    assert np.array_equal(result, np.arange(len(gdf)))
+
+
+def test_get_id_col_array_uses_index():
+    """_get_id_col_array should use the GeoDataFrame index when requested."""
+    gdf = gpd.GeoDataFrame({"id": [10, 11], "geometry": [Point(0, 0), Point(1, 1)]})
+    gdf.index = ["a", "b"]
+
+    result = _get_id_col_array(gdf, "index")
+
+    assert np.array_equal(result, gdf.index.values)
+
+
+def test_get_id_col_array_uses_column():
+    """_get_id_col_array should use a named column."""
+    gdf = gpd.GeoDataFrame({"id": [10, 11], "geometry": [Point(0, 0), Point(1, 1)]})
+
+    result = _get_id_col_array(gdf, "id")
+
+    assert np.array_equal(result, gdf["id"].values)
+
+
+def test_get_id_col_array_duplicate_values_error():
+    """Duplicate ids should raise a ValueError."""
+    gdf = gpd.GeoDataFrame({"id": [10, 10], "geometry": [Point(0, 0), Point(1, 1)]})
+
+    with pytest.raises(ValueError, match="ID column must contain unique values"):
+        _get_id_col_array(gdf, "id")
+
+
+def test_get_id_col_array_multiindex_error():
+    """MultiIndex should raise a ValueError when using index ids."""
+    gdf = gpd.GeoDataFrame({"id": [10, 11], "geometry": [Point(0, 0), Point(1, 1)]})
+    gdf.index = pd.MultiIndex.from_tuples([(0, "a"), (1, "b")])
+
+    with pytest.raises(ValueError, match="Multi-index is not supported"):
+        _get_id_col_array(gdf, "index")
+
+
 def test_id_col_not_duplicated_as_node_attribute_when_used_as_node_key():
     """Test that the id column is not duplicated as a node attribute when used as a node key."""
     gdf = gpd.GeoDataFrame(
@@ -88,3 +141,212 @@ def test_id_col_not_duplicated_as_node_attribute_when_used_as_node_key():
 
     # check an example value matches
     assert graph.nodes[101]["foo"] == "a"
+
+
+def test_graph_to_gdf_requires_layer_selection():
+    """nodes=False and edges=False should raise."""
+    g = nx.Graph()
+
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
[…]
