### 2026-02-11 10b65b6
update version number to 0.4.0
