+++ pysgn/__init__.py
@@ -27,7 +27,7 @@ __all__ = [
 ]
 
 __title__ = "pysgn"
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 __license__ = "MIT"
 _this_year = datetime.datetime.now(tz=datetime.timezone.utc).date().year
 __copyright__ = f"Copyright {_this_year} Wang Boyu"
