### 2026-02-12 #9346 @tomalrussell
Pending merge and release of an updated version for most remaining points in my review.

Otherwise as noted, the outstanding point is that this package is on the light side of the "scholarly effort" criterion, but within reach. In particular "feature complete" is hard to judge, but the synthesis of complete spatial networks from scratch would seem desirable to me, not only edge inference/generation. I'd encourage some more development effort within the scope of this review.
