### 2026-02-13 #9346 @wang-boyu
Thank you @tomalrussell and @jGaboardi. I have merged and released most of my open PRs. The only one remains open is blocked by another PR I submitted to conda-forge to add my package into it. There're hundreds of open PR with them so I don't expect mine to be merged anytime soon.

In terms of the feature request for node synthesis, I'll talk to my co-authors and discuss further. Updates will be posted in the issue ticket once there's any.
