### 2026-02-13 #9346 @wang-boyu
Absolutely. Here it is: https://github.com/conda-forge/staged-recipes/pull/31762
