### 2026-02-13 1b2cfd2
fix latex equation in docstrings
