+++ pysgn/geo_erdos_renyi_network.py
@@ -34,9 +34,9 @@ def geo_erdos_renyi_network(
     Each possible edge in the network is connected with probability:
 
     .. math::
-        p(\textrm{distance}|a, \textrm{min_dist}) = \textrm{min}\left(1, \left(\frac{\textrm{distance}}{\textrm{min_dist}}\right) ^ {-a}\right)
+        p(\textrm{distance}|a, \textrm{min\_dist}) = \textrm{min}\left(1, \left(\frac{\textrm{distance}}{\textrm{min\_dist}}\right) ^ {-a}\right)
 
-    where min_dist is the minimum distance between nodes, and a is the distance decay exponent parameter, default is 3.
+    where :math:`min\_dist` is the minimum distance between nodes, and :math:`a` is the distance decay exponent parameter, default is 3.
     The minimum distance is a threshold, below which nodes are connected with probability 1, if an edge is chosen to be rewired.
     It is 1/20 of the bounding box diagonal by default. Users can set the scaling factor directly if needed, which is the inverse of the minimum distance.
 
