### 2026-02-15 49b25d3
add sample_points function for node synthesis
