+++ pysgn/utils.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
 import warnings
+from collections.abc import Callable, Sequence
+from numbers import Real
 from typing import Any
 
 import geopandas as gpd
@@ -8,8 +10,213 @@ import networkx as nx
 import numpy as np
 import pandas as pd
 from loguru import logger
-from shapely.geometry import LineString, Point
+from pyproj import CRS
+from shapely.geometry import LineString, MultiPolygon, Point, Polygon
 from shapely.geometry.base import BaseGeometry
+
+
+def sample_points(
+    n: int,
+    *,
+    bbox: Sequence[float] | None = None,
+    polygon: Polygon | MultiPolygon | None = None,
+    sampler: Callable[[np.random.Generator, int], np.ndarray] | None = None,
+    random_state: int | None = None,
+    crs: Any | None = None,
+    max_attempts: int = 1_000_000,
+) -> gpd.GeoDataFrame:
+    """Sample point geometries within a spatial domain.
+
+    Returns a geometry-only ``GeoDataFrame`` with exactly ``n`` point
+    geometries in the ``geometry`` column.
+
+    Domain definition:
+
+    - ``bbox`` only: sample in ``(xmin, ymin, xmax, ymax)`` bounds.
+    - ``polygon`` only: sample in polygon area.
+    - both: sample in ``polygon.intersection(bbox_polygon)``.
+    - at least one of ``bbox`` or ``polygon`` must be provided.
+
+    Sampling:
+
+    - ``sampler is None``: uniform candidate generation over the final domain.
+    - ``sampler is callable``: user-defined candidate generation with
+      ``sampler(rng, k) -> array-like`` of shape ``(k, 2)``, followed by
+      rejection against the final domain.
+
+    CRS handling:
+
+    - CRS is set only from explicit ``crs`` input with ``pyproj.CRS.from_user_input``.
+
+    Args:
+        n: Number of points to generate; must be > 0.
+        bbox: Bounding box sequence ``(xmin, ymin, xmax, ymax)``.
+        polygon: Domain polygon geometry (``Polygon`` or ``MultiPolygon`` only).
+        sampler: Optional callable producing candidate coordinates.
+        random_state: Random seed for deterministic output.
+        crs: Optional CRS user input passed through
+            ``pyproj.CRS.from_user_input`` before assignment.
+        max_attempts: Maximum number of evaluated candidate points.
+
+    Returns:
+        A geometry-only ``geopandas.GeoDataFrame`` with ``n`` sampled points.
+
+    Raises:
+        ValueError: For invalid inputs (for example ``n <= 0``, invalid
+            ``bbox``/``polygon``, missing domain, invalid ``sampler``,
+            ``max_attempts <= 0``, or non-positive final domain area).
+        RuntimeError: If sampling cannot produce ``n`` accepted points within
+            ``max_attempts`` evaluated candidate points.
+    """
+    if not isinstance(n, (int, np.integer)) or n <= 0:
+        raise ValueError("n must be a positive integer.")
+
+    if not isinstance(max_attempts, (int, np.integer)) or max_attempts <= 0:
+        raise ValueError("max_attempts must be a positive integer.")
+
+    if sampler is not None and not callable(sampler):
+        raise ValueError("sampler must be callable or None.")
+
+    final_domain = _resolve_sampling_domain(bbox=bbox, polygon=polygon)
+    rng = np.random.default_rng(seed=random_state)
+    parsed_crs = CRS.from_user_input(crs) if crs is not None else None
+
+    accepted_coords: list[tuple[float, float]] = []
+    attempts = 0
+
+    while len(accepted_coords) < n and attempts < max_attempts:
[…]
