### 2026-02-15 86ca5d4
add sample_points function for node synthesis
