+++ pysgn/utils.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
 import warnings
+from collections.abc import Callable, Sequence
+from numbers import Real
 from typing import Any
 
 import geopandas as gpd
@@ -8,8 +10,220 @@ import networkx as nx
 import numpy as np
 import pandas as pd
 from loguru import logger
-from shapely.geometry import LineString, Point
+from pyproj import CRS
+from shapely.geometry import LineString, MultiPolygon, Point, Polygon
 from shapely.geometry.base import BaseGeometry
+
+
+def sample_points(
+    n: int,
+    *,
+    bbox: Sequence[float] | None = None,
+    polygon: Polygon | MultiPolygon | None = None,
+    sampler: Callable[[np.random.Generator, int], np.ndarray] | None = None,
+    random_state: int | None = None,
+    crs: Any | None = None,
+    max_attempts: int = 1_000_000,
+) -> gpd.GeoDataFrame:
+    """Sample point geometries within a spatial domain.
+
+    Phase 1 API for node synthesis. This function returns a geometry-only
+    ``GeoDataFrame`` with exactly ``n`` point geometries in the ``geometry``
+    column.
+
+    Domain definition:
+    - ``bbox`` only: sample in ``(xmin, ymin, xmax, ymax)`` bounds
+    - ``polygon`` only: sample in polygon area
+    - both: sample in ``polygon.intersection(bbox_polygon)``
+    - at least one of ``bbox`` or ``polygon`` must be provided
+
+    Sampling modes:
+    - ``sampler is None``: uniform candidate generation over the final domain
+    - ``sampler is callable``: user-driven candidate generation using the
+      contract ``sampler(rng, k) -> array-like`` with exact shape ``(k, 2)``,
+      numeric finite values, then rejection against the final domain
+
+    Determinism:
+    - ``random_state`` may be an ``int`` or ``None``
+    - RNG initialization uses ``np.random.default_rng(seed=random_state)``
+
+    CRS:
+    - CRS is set only from explicit ``crs``
+    - explicit CRS values are normalized with
+      ``pyproj.CRS.from_user_input(crs)``
+      (https://pyproj4.github.io/pyproj/stable/api/crs/crs.html#pyproj.crs.CRS.from_user_input)
+    - no CRS inference is performed from ``polygon``
+
+    Error behavior:
+    - Raises ``ValueError`` for invalid inputs (for example: ``n <= 0``,
+      invalid ``bbox``, invalid/empty ``polygon``, missing domain, non-callable
+      ``sampler``, invalid sampler output, ``max_attempts <= 0``, or final
+      domain area ``<= 0``)
+    - Raises ``RuntimeError`` if sampling cannot produce ``n`` accepted points
+      within ``max_attempts`` evaluated candidate points
+
+    Args:
+        n: Number of points to generate; must be > 0.
+        bbox: Bounding box sequence ``(xmin, ymin, xmax, ymax)``.
+        polygon: Domain polygon geometry (``Polygon`` or ``MultiPolygon`` only).
+        sampler: Optional callable producing candidate coordinates.
+        random_state: Random seed for deterministic output.
+        crs: Optional CRS user input passed through
+            ``pyproj.CRS.from_user_input`` before assignment.
+        max_attempts: Maximum number of evaluated candidate points.
+
+    Returns:
+        A geometry-only ``geopandas.GeoDataFrame`` with ``n`` sampled points.
+    """
+    if not isinstance(n, (int, np.integer)) or n <= 0:
+        raise ValueError("n must be a positive integer.")
+
+    if not isinstance(max_attempts, (int, np.integer)) or max_attempts <= 0:
+        raise ValueError("max_attempts must be a positive integer.")
+
[…]
