+++ tests/test_utils.py
@@ -1,17 +1,22 @@
 import re
+from functools import partial
 
 import geopandas as gpd
 import networkx as nx
 import numpy as np
 import pandas as pd
 import pytest
-from shapely.geometry import LineString, Point, Polygon
+from shapely.geometry import LineString, MultiPolygon, Point, Polygon
 
 from pysgn.utils import (
+    _bbox_to_polygon,
     _create_k_col,
     _get_id_col_array,
+    _resolve_sampling_domain,
     _set_node_attributes,
+    _validate_polygon,
     graph_to_gdf,
+    sample_points,
 )
 
 
@@ -66,6 +71,358 @@ def test_create_k_col_reproducibility():
     assert np.array_equal(result1, result2), (
         "Results are not reproducible with the same random state"
     )
+
+
+def test_bbox_to_polygon_valid_bbox_returns_rectangle():
+    """_bbox_to_polygon should return a rectangle with expected bounds."""
+    poly = _bbox_to_polygon((0, 1, 3, 5))
+
+    assert isinstance(poly, Polygon)
+    assert poly.bounds == (0.0, 1.0, 3.0, 5.0)
+    assert poly.area == 12.0
+
+
+def test_bbox_to_polygon_accepts_numeric_sequence_types():
+    """_bbox_to_polygon should accept mixed real numeric sequence inputs."""
+    poly = _bbox_to_polygon(np.array([0, 0.5, 2, 1.5]))
+
+    assert poly.bounds == (0.0, 0.5, 2.0, 1.5)
+
+
+@pytest.mark.parametrize("bbox", [None, 5, object()])
+def test_bbox_to_polygon_rejects_non_sequence_input(bbox):
+    """_bbox_to_polygon should reject non-sequence bbox values."""
+    with pytest.raises(
+        ValueError,
+        match="bbox must be a sequence of 4 values",
+    ):
+        _bbox_to_polygon(bbox)  # type: ignore[arg-type]
+
+
+@pytest.mark.parametrize("bbox", [(0, 1, 2), (0, 1, 2, 3, 4)])
+def test_bbox_to_polygon_rejects_wrong_length(bbox):
+    """_bbox_to_polygon should reject bbox with incorrect length."""
+    with pytest.raises(
+        ValueError,
+        match="bbox must be a sequence of 4 values",
+    ):
+        _bbox_to_polygon(bbox)
+
+
+def test_bbox_to_polygon_rejects_non_numeric_values():
+    """_bbox_to_polygon should reject non-numeric bbox entries."""
+    with pytest.raises(ValueError, match="bbox values must be numeric"):
+        _bbox_to_polygon((0, "x", 2, 3))  # type: ignore[arg-type]
+
+
+@pytest.mark.parametrize("bad_value", [np.nan, np.inf, -np.inf])
+def test_bbox_to_polygon_rejects_non_finite_values(bad_value):
+    """_bbox_to_polygon should reject NaN/Inf values."""
+    with pytest.raises(ValueError, match="bbox values must be finite"):
+        _bbox_to_polygon((0, 1, bad_value, 4))
+
+
+@pytest.mark.parametrize(
+    "bbox", [(0, 0, 0, 1), (0, 2, 1, 2), (2, 0, 1, 1), (0, 3, 1, 2)]
+)
+def test_bbox_to_polygon_rejects_invalid_ordering(bbox):
+    """_bbox_to_polygon should enforce xmin<xmax and ymin<ymax."""
+    with pytest.raises(
+        ValueError,
+        match="bbox ordering must satisfy xmin < xmax and ymin < ymax",
+    ):
+        _bbox_to_polygon(bbox)
+
+
[…]
