### 2026-02-15 4778260
wip: refactor rtd docs
