+++ docs/conf.py
@@ -86,7 +86,7 @@ release = "0.4.1"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = ["_build"]
+exclude_patterns = ["_build", "jupyter_execute"]
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
