### 2026-02-15 6eb44ea
remove redundant pos column from exported node gdf
