### 2026-02-20 #9346 @jGaboardi
* @tomalrussell – With latest release of `pysgn` now available via [`conda-forge`](https://anaconda.org/channels/conda-forge/packages/pysgn/overview), I think that most of your boxes can be check out (thought perhaps not all?) When you have a moment, can you update those, and then comment down here the remaining concerns you have.
* @gegen07 – Please provide your comment for @wang-boyu most recent updates as time permits.
