### 2026-02-22 #9346 @gegen07
Hi @jGaboardi and @wang-boyu, all tickets are cleared.  Although PySGN does not provide an innovative method for random graph generation, it is a geospatial extension of NetworkX methods, thus being a good complement for geospatial environment packages.

I would recommend acceptance after solving Tom's remarks.
