### 2026-02-22 #9346 @jGaboardi
> I would recommend acceptance after solving Tom's remarks.

@gegen07 – Thank you for your update. I can see that you have 6 empty boxes in your [review checklist](https://github.com/openjournals/joss-reviews/issues/9346#issuecomment-3506853619). Is there any preventing those boxes from being checked or are tied to the resolution of @tomalrussell's comments? (not trying to pressure you at all, simply would like to clarify).
