### 2026-02-22 #9346 @jGaboardi
#### Note to Managing EiC

@samhforbes – While this review is still underway, I wanted to let you know that during the review process several works that I was directly involved with[^1][^2] were added to the references. I wanted to bring this to your attention to head off any concerns of impropriety that may arise later.

[^1]: [Gaboardi, J. D., Rey, S., & Lumnitz, S., (2021). spaghetti: spatial network analysis in PySAL. Journal of Open Source Software, 6(62), 2826](https://joss.theoj.org/papers/10.21105/joss.02826)
[^2]: [Fleischmann, M., Vybornova, A., Gaboardi, J. D., Brázdová, A., & Dančejová, D. (2026). Adaptive continuity-preserving simplification of street networks. Computers, Environment and Urban Systems, 123, 102354.](https://doi.org/10.1016/j.compenvurbsys.2025.102354)
