### 2026-02-22 #9346 @jGaboardi
@wang-boyu – Please take this time to review your complete reference list for proper casing in all entries (e.g.,`(Züfle et al., 2024)`'s title [should be capitalized](https://dl.acm.org/doi/10.1145/3672557#:~:text=Share%20on-,In%20Silico%20Human%20Mobility%20Data%20Science%3A%20Leveraging%20Massive%20Simulated%20Mobility%20Data,-(Vision%20Paper))), DOI inclusion (e.g., `(Rey & Anselin, 2007)` is missing [it's DOI](https://www.scopus.com/pages/publications/67349102538#:~:text=DOI:%2010.52324/001c.8285)) etc.
