### 2026-02-22 c6bb5fe
update references in paper
