+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @inproceedings{10.1145/3615896.3628345,
   author = {Gallagher, Ketevan and Anderson, Taylor and Crooks, Andrew and Z\"{u}fle, Andreas},
-  title = {Synthetic Geosocial Network Generation},
+  title = {{Synthetic Geosocial Network Generation}},
   year = {2023},
   isbn = {9798400703584},
   publisher = {Association for Computing Machinery},
@@ -17,7 +17,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
   }
 
   @inbook{erdos1960evolution,
-  title = {On the Evolution of Random Graphs},
+  title = {On the evolution of random graphs},
   author = {P. Erd\"os and A. R\'enyi},
   journal={Publ. math. inst. hung. acad. sci},
   volume={5},
@@ -27,7 +27,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
   }
 
 @article{watts1998collective,
-  title={Collective Dynamics of 'Small-World' Networks},
+  title={Collective dynamics of 'small-world' networks},
   author={Watts, Duncan J and Strogatz, Steven H},
   journal={nature},
   volume={393},
@@ -39,7 +39,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
 }
 
 @article{alizadeh2017generating,
-  title={Generating and Analyzing Spatial Social Networks},
+  title={Generating and analyzing spatial social networks},
   author={Alizadeh, Meysam and Cioffi-Revilla, Claudio and Crooks, Andrew},
   journal={Computational and Mathematical Organization Theory},
   volume={23},
@@ -51,7 +51,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
 
 @conference{osti_960616,
   author       = {Hagberg, Aric and Swart, Pieter J. and Schult, Daniel A.},
-  title        = {Exploring Network Structure, Dynamics, and Function Using NetworkX},
+  title        = {{Exploring Network Structure, Dynamics, and Function {using} {NetworkX}}},
   url          = {https://www.osti.gov/biblio/960616},
   place        = {United States},
   organization = {Los Alamos National Laboratory (LANL)},
@@ -90,7 +90,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
                   John Flavin and
                   Sergio Rey and
                   James Gardiner},
-  title        = {geopandas/geopandas: v1.0.1},
+  title        = {{geopandas}/geopandas: v1.0.1},
   month        = jul,
   year         = 2024,
   publisher    = {Zenodo},
@@ -100,7 +100,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
 }
 
 @article{jiang2024large,
-  title={A Large-Scale Geographically Explicit Synthetic Population with Social Networks for the United States},
+  title={{A {Large-Scale} Geographically Explicit Synthetic Population with Social Networks for the United States}},
   author={Jiang, Na and Yin, Fuzhen and Wang, Boyu and Crooks, Andrew T},
   journal={Scientific Data},
   volume={11},
@@ -112,7 +112,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
 }
 
 @article{barabasi1999emergence,
-  title={Emergence of Scaling in Random Networks},
+  title={{Emergence of Scaling in Random Networks}},
   author={Barab{\'a}si, Albert-L{\'a}szl{\'o} and Albert, R{\'e}ka},
   journal={Science},
   volume={286},
[…]
