### 2026-02-22 #9346 @wang-boyu
Hi @jGaboardi I'm checking the title casing for references, and it's a mix of title case and sentence case if we follow the capitalization used by the paper authors. Do we use these, or do you prefer it's all converted to title / sentence case?
