### 2026-02-22 #9346 @jGaboardi
> Hi [@jGaboardi](https://github.com/jGaboardi) I'm checking the title casing for references, and it's a mix of title case and sentence case if we follow the capitalization used by the paper authors. Do we use these, or do you prefer it's all converted to title / sentence case?

I prefer to stick with the case used in each reference's official citation. So it will be a mix (like you point out).
