### 2026-02-22 #9346 @wang-boyu
Done! I've updated the title casing and doi for the references. Please let me know if there's any issues with it @jGaboardi.
