### 2026-02-22 fc27889
Update paper.bib
