+++ paper/paper.bib
@@ -7,19 +7,19 @@
   address = {New York, NY, USA},
   url = {https://doi.org/10.1145/3615896.3628345},
   doi = {10.1145/3615896.3628345},
-  booktitle = {Proceedings of the 7th ACM SIGSPATIAL Workshop on Location-Based Recommendations, Geosocial Networks and Geoadvertising},
-  pages = {15–24},
+  booktitle = {Proceedings of the 7\textsuperscript{th} ACM SIGSPATIAL Workshop on Location-Based Recommendations, Geosocial Networks and Geoadvertising},
+  pages = {15--24},
   numpages = {10},
   keywords = {geosocial networks, network generation, synthetic social
 networks, erdos-renyi, watts-strogatz, barabasi-albert},
   location = {Hamburg, Germany},
-  series = {LocalRec '23}
+  series = {LocalRec `23}
   }
 
   @inbook{erdos1960evolution,
   title = {On the evolution of random graphs},
   author = {P. Erd\"os and A. R\'enyi},
-  journal={Publ. math. inst. hung. acad. sci},
+  journal={{Publ. Math. Inst. Hungar. Acad. Sci}},
   volume={5},
   number={1},
   pages={17--60},
@@ -27,7 +27,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
   }
 
 @article{watts1998collective,
-  title={Collective dynamics of 'small-world' networks},
+  title={Collective dynamics of `small-world' networks},
   author={Watts, Duncan J and Strogatz, Steven H},
   journal={nature},
   volume={393},
@@ -179,7 +179,7 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
   year=2007,
   volume={37},
   number={1},
-  pages={5-27},
+  pages={5--27},
   keywords={Open Source; Software; Spatial},
   doi={10.52324/001c.8285}
 }
@@ -190,8 +190,10 @@ networks, erdos-renyi, watts-strogatz, barabasi-albert},
   journal={InterJournal Complex Systems},
   volume={1695},
   pages={1--9},
-  year={2006}
+  year={2006},
+  url={https://igraph.org},
 }
+
 @Article{Gastner2006,
 author={Gastner, M. T.
 and Newman, M. E. J.},
