### 2026-03-06 #9346 @jGaboardi
@gegen07 – All of [Tom's boxes are now checked off](https://github.com/openjournals/joss-reviews/issues/9346#issuecomment-3492621888), if you have no further comments please check off your boxes (if you are satisfied). Once that is complete, I will start my pre-acceptance editor's pass.
