### 2026-03-06 #9346 @jGaboardi
@wang-boyu  – please review the *Additional Author Tasks After Review is Complete* [above](https://github.com/openjournals/joss-reviews/issues/9346#issuecomment-4009208369) for those 5 bullet points.

Once you have completed those bullets, ping me here and I will start on the *Editor Tasks Prior to Acceptance*.
