### 2026-03-06 ebe6b82
update version number to 0.5.0
