### 2026-03-06 #9346 @wang-boyu
Thans again @tomalrussell and @gegen07 for taking your time in reviewing our paper and providing your comments. I'll close most of the tracking issues in the repository as completed, but please feel free to re-open if you have any additional suggestions. Some features are considered as nice to have, and I'll leave them open for future updates.

Hi @jGaboardi regarding the Additional Author Tasks:

- Version number with latest changes from the review: `0.5.0`
- DOI from Zenodo archive: 10.5281/zenodo.18883898
