### 2026-03-07 #9346 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
