### 2026-03-15 09c3135
add citation and update version number
