+++ docs/conf.py
@@ -70,9 +70,9 @@ copyright = f"2017-{date.today().year}, Wang Boyu"
 # built documents.
 #
 # The short X.Y version.
-version = "0.5.0"
+version = "0.5.1"
 # The full version, including alpha/beta/rc tags.
-release = "0.5.0"
+release = "0.5.1"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
